Added positioning system, based on pathing system.
This also has some other junk rolled into it, like better debugging for paths and some pretty good debugging for seeing where the player is projected onto the path. Some things that need to be changed:
There is a bug that might leave a player's "position" back a couple checkpoints. I'll have to consider how to fix it. Also, some of this debug stuff should be removed at some point; The overlay and the debug messages, specifically. The debug line draws shouldn't appear in production, but they could be removed to.