Rewrite dumb position tracking code
The previous code was convoluted and had a bug where the player wasn't really projected onto the nearest curve in all cases; now it should work in general. It'll at least work in more cases than it did before. Also, this code is a lot clearer in what it's doing than the other code.