Fixed boost -> back boost bug
Before, when a player would then hit a boost, then hit a backwards boost, they would have a max speed of about 180 when they finally got going forward again. I added a couple of lines to STATES.BOOST and STATES.BOOST_ B that reset the maxSpeed in STATES.BOOST_ B or the maxReverse in STATES.BOOST to what they are supposed to be.